Fix max length for German numbers to 11
Reference:

"A new numbering plan was introduced on 3 May 2010. Since then new landline
phone numbers have a standard length of 11 digits, which includes the
area code but omits the trunk prefix of 0. Area codes remain as they
are and are still variable in length. Exceptions to the 11 digit rule
are the four cities of Berlin, Frankfurt, Hamburg and Munich, which are
the only cities with two digit area codes and require only 10 digit
numbers so as not to exceed the maximum length of 8 digits for a
subscriber number."

https://en.wikipedia.org/wiki/Telephone_numbers_in_Germany